Moved artifact tool and related modules up to common.
