--- azure-devops/azext_devops/dev/artifacts/const.py
@@ -1,8 +1,0 @@
-# artifact constants
-from azext_devops.dev.common.const import CLI_ENV_VARIABLE_PREFIX
-
-
-ARTIFACTTOOL_OVERRIDE_PATH_ENVKEY = CLI_ENV_VARIABLE_PREFIX + "ARTIFACTTOOL_OVERRIDE_PATH"
-ARTIFACTTOOL_OVERRIDE_URL_ENVKEY = CLI_ENV_VARIABLE_PREFIX + "ARTIFACTTOOL_OVERRIDE_URL"
-ARTIFACTTOOL_OVERRIDE_VERSION_ENVKEY = CLI_ENV_VARIABLE_PREFIX + "ARTIFACTTOOL_OVERRIDE_VERSION"
-ARTIFACTTOOL_PAT_ENVKEY = CLI_ENV_VARIABLE_PREFIX + "ARTIFACTTOOL_PATVAR"
