+++ azure-devops/azext_devops/dev/common/const.py
@@ -24,3 +24,35 @@ DEVOPS_ORGANIZATION_DEFAULT = 'organization'
 DEVOPS_TEAM_PROJECT_DEFAULT = 'project'
 PAT_ENV_VARIABLE_NAME = CLI_ENV_VARIABLE_PREFIX + 'PAT'
 AUTH_TOKEN_ENV_VARIABLE_NAME = CLI_ENV_VARIABLE_PREFIX + 'AUTH_TOKEN'
+
+# policy constants
+
+APPROVER_COUNT_POLICY = 'ApproverCountPolicy'
+APPROVER_COUNT_POLICY_ID = 'fa4e907d-c16b-4a4c-9dfa-4906e5d171dd'
+
+BUILD_POLICY = 'BuildPolicy'
+BUILD_POLICY_ID = '0609b952-1397-4640-95ec-e00a01b2c241'
+
+COMMENT_REQUIREMENTS_POLICY = 'CommentRequirementsPolicy'
+COMMENT_REQUIREMENTS_POLICY_ID = 'c6a1889d-b943-4856-b76f-9e46bb6b0df2'
+
+MERGE_STRATEGY_POLICY = 'MergeStrategyPolicy'
+MERGE_STRATEGY_POLICY_ID = 'fa4e907d-c16b-4a4c-9dfa-4916e5d171ab'
+
+FILE_SIZE_POLICY = 'FileSizePolicy'
+FILE_SIZE_POLICY_ID = '2e26e725-8201-4edd-8bf5-978563c34a80'
+
+WORKITEM_LINKING_POLICY = 'WorkItemLinkingPolicy'
+WORKITEM_LINKING_POLICY_ID = '40e92b44-2fe1-4dd6-b3d8-74a9c21d0c6e'
+
+REQUIRED_REVIEWER_POLICY = 'RequiredReviewersPolicy'
+REQUIRED_REVIEWER_POLICY_ID = 'fd2167ab-b0be-447a-8ec8-39368250530e'
+
+REPO_POLICY_TYPE = [APPROVER_COUNT_POLICY, BUILD_POLICY, COMMENT_REQUIREMENTS_POLICY, MERGE_STRATEGY_POLICY,
+                    FILE_SIZE_POLICY, WORKITEM_LINKING_POLICY, REQUIRED_REVIEWER_POLICY]
+
+# Artifact tool variables
+ARTIFACTTOOL_OVERRIDE_PATH_ENVKEY = CLI_ENV_VARIABLE_PREFIX + "ARTIFACTTOOL_OVERRIDE_PATH"
+ARTIFACTTOOL_OVERRIDE_URL_ENVKEY = CLI_ENV_VARIABLE_PREFIX + "ARTIFACTTOOL_OVERRIDE_URL"
+ARTIFACTTOOL_OVERRIDE_VERSION_ENVKEY = CLI_ENV_VARIABLE_PREFIX + "ARTIFACTTOOL_OVERRIDE_VERSION"
+ARTIFACTTOOL_PAT_ENVKEY = CLI_ENV_VARIABLE_PREFIX + "ARTIFACTTOOL_PATVAR"
